api/requirements.txt
pydantic-ai==0.0.14
fastapi==0.115.6
uvicorn[standard]==0.34.0
google-generativeai==0.8.3
python-dotenv==1.0.1
pydantic==2.10.3
httpx==0.28.1
